One more int-->uint8_t

+++ src/dials/array_family/shoebox_extractor.h
@@ -121,7 +121,7 @@ namespace dials { namespace af {
     void next(const Image<T>& image) {
       typedef Shoebox<>::float_type float_type;
       typedef af::ref<float_type, af::c_grid<3> > sbox_data_type;
-      typedef af::ref<int, af::c_grid<3> > sbox_mask_type;
+      typedef af::ref<uint8_t, af::c_grid<3> > sbox_mask_type;
       DIALS_ASSERT(frame_ >= frame0_ && frame_ < frame1_);
       DIALS_ASSERT(image.npanels() == npanels_);
       for (std::size_t p = 0; p < image.npanels(); ++p) {
